requirements.txt
backports.functools-lru-cache==1.5
bitstruct==3.10.0
cantools==28.7.2
cycler==0.10.0
kiwisolver==1.0.1
numpy==1.15.1
protobuf==3.6.1
pyparsing==2.2.0
python-can==2.2.1
python-dateutil==2.7.3
pytz==2018.5
pyzmq==17.1.2
six==1.11.0
subprocess32==3.5.2
textparser==0.17.0
wrapt==1.10.11
udsoncan==1.3
can-isotp==0.4